production_app/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
torch
torchvision
numpy
pillow
scikit-image
ultralytics
transformers
kagglehub
requests
